requirements.txt
django
djangorestframework
django-cors-headers
langchain
langchain-core
langchain-openai
requests
